refactor(storage): split index.ts into focused modules
index.ts drops from 1843 to 1551 LOC. Four new files, each named for
one concern:

  src/driver.ts             driver types + assertMutable
  src/migrations.ts         defineMigrations, DefineMigrations, MigrationError
  src/pipeline.ts           runReadPipeline, runWritePipeline, adapter types,
                            defineSchema, PipelineValidateResult
  src/pipeline-adapters.ts  processReadValue, processWriteValue as pure
                            module-level functions taking driver/key/opts

Public API unchanged; every export from index.ts is preserved via
re-export from the new modules where relocated. 171/171 tests pass,
TypeScript clean under all strict flags.

Type-safety wins from the refactor:
- PipelineValidateResult<TOut> is now StandardSchemaV1.Result<TOut>,
  the same discriminated union Standard Schema itself uses. The old
  inline pattern in processReadValue / processWriteValue required TS
  to work around a non-discriminated shape; the extracted pipeline
  narrows cleanly on `if (result.issues)`, so `result.value` types
  as TOut with no cast.
- The adapter shape (hasSchema, passthrough, preValidate, validate,
  fallback, onValidationError, resetUnderLock) is a single place to
  reason about the read pipeline. Encoder-at-rest and other future
  serializers plug in via the adapter without touching pipeline code.
- Zero `as any` / `as never` / `as unknown as` escapes in the new
  files. Every generic in the pipeline satisfies the Golden Rule
  (each type parameter appears \u2265 2 times in the signature).

No behavioural change: the four extracted functions retain their
original semantics; only their location changes.
